+++ CHANGES.md
@@ -1,4 +1,4 @@
-* @dev
+* v1.3.0 (2020-12-18)
   * Added update availability checker
   * Corrected layout items